+++ src/main/java/com/untels/controller/admin/CategoriaController.java
@@ -1,7 +1,12 @@
 package com.untels.controller.admin;
 
 import com.untels.service.CategoriaService;
+
+import javax.servlet.http.HttpSession;
+
 import com.untels.entity.Categoria;
+import com.untels.enums.Rol;
+import com.untels.security.supervisor.Supervisor;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,14 +21,35 @@ public class CategoriaController {
     @Autowired
     CategoriaService categoriaService;
 
+    @Autowired
+    HttpSession session;
+
     @GetMapping("/admin/categorias")
     public String gestionCategorias(Model model) {
+
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
+
         model.addAttribute("listaCategorias", categoriaService.findAll());
         return "admin/categorias/lista-categorias";
     }
 
     @GetMapping("/admin/categorias/nuevo")
     public String crearCategoria(Model model) {
+
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
+      
         Categoria categoria = new Categoria();
         model.addAttribute("categoria", categoria);
         return "admin/categorias/form-categoria";
@@ -31,6 +57,14 @@ public class CategoriaController {
 
     @PostMapping("/admin/categorias/grabar")
     public String grabarCategoria(@ModelAttribute("categoria") Categoria categoriaParam) {
+
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
 
         Categoria categoria = new Categoria();
 
@@ -46,6 +80,14 @@ public class CategoriaController {
     @GetMapping("/admin/categorias/{id}")
     public String editarCategoria(@PathVariable("id") int id, Model model) {
 
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
+
         if (!categoriaService.existsByIdCategoria(id)) {
             // TODO: Cambiar a pagina de error
             return "pagina-404";
@@ -57,6 +99,14 @@ public class CategoriaController {
 
     @PostMapping("/admin/categorias/editar")
     public String editarCategoriaForm(@ModelAttribute("categoria") Categoria categoriaParam) {
+
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
 
         Categoria categoria = categoriaService.findByIdCategoria(categoriaParam.getIdCategoria());
 
@@ -71,6 +121,15 @@ public class CategoriaController {
 
     @GetMapping("/admin/categorias/eliminar/{id}")
     public String eliminarCategoria(@PathVariable("id") int id) {
+      
+        if (!Supervisor.haIniciadoSesion(session)) {
+            return "redirect:/admin";
+        }
+
+        if (!Supervisor.tieneRol(session, Rol.ADMIN)) {
+            return "redirect:/admin";
+        }
+      
         categoriaService.deleteByIdCategoria(id);
         return  "redirect:/admin/categorias";
     }
